Update complex ip logger python
This is an ip logger that sends someone's ip to a discord server. It checks their ip every three seconds therefore it will automatically send the new ip address in case they have changed it.
Thus if the person is using a VPN and happens to turn it off for at least 1 second it will detect the original ip address and send it to the server as well. 
This code also sends your computer info, for example, your system into the server.